requirements.txt
fastapi
pycryptodome
phpserialize3
python-dotenv